* commit 'b0cbf8607fb00a2bc54d15e01570cabf147fb5fb':
  Tracking libcore change to use longs instead of ints to store pointers in OpenSSL binding

+++ core/java/android/webkit/ClientCertRequestHandler.java
@@ -75,7 +75,7 @@ public final class ClientCertRequestHandler extends Handler {
     /**
      * Proceed with the specified private key bytes and client certificate chain.
      */
-    private void setSslClientCertFromCtx(final int ctx, final byte[][] chainBytes) {
+    private void setSslClientCertFromCtx(final long ctx, final byte[][] chainBytes) {
         post(new Runnable() {
                 public void run() {
                     mBrowserFrame.nativeSslClientCert(mHandle, ctx, chainBytes);
